Flex: add types file
