+++ src/Layout/Flex/index.tsx
@@ -2,29 +2,7 @@ import * as React from 'react';
 import styled from 'styled-components';
 
 import { Box, Props as BoxProps } from '../Box';
-
-const positions = ['center', 'flex-start', 'flex-end'] as const;
-
-export const FlexPropAndPossibleValueMap = {
-  flexDirection: ['row', 'row-reverse', 'column', 'column-reverse'],
-  justifyContent: [...positions, 'space-between', 'space-around'],
-  alignItems: [...positions, 'stretch', 'baseline'] as const,
-  flexWrap: ['nowrap', 'wrap', 'wrap-reverse'] as const,
-};
-
-export const FlexPropAndDefaultValueMap: FlexProps = {
-  flexDirection: 'row',
-  justifyContent: 'flex-start',
-  alignItems: 'stretch',
-  flexWrap: 'nowrap',
-};
-
-type FlexProps = {
-  flexDirection?: typeof FlexPropAndPossibleValueMap.flexDirection[number];
-  justifyContent?: typeof FlexPropAndPossibleValueMap.justifyContent[number];
-  alignItems?: typeof FlexPropAndPossibleValueMap.alignItems[number];
-  flexWrap?: typeof FlexPropAndPossibleValueMap.flexWrap[number];
-};
+import { FlexProps, FlexPropAndDefaultValueMap } from './types';
 
 type Props = FlexProps & BoxProps;
